refactor locations controller to be generaic
